LayoutTests/http/tests/security/contentSecurityPolicy/1.1/module-scriptnonce-ignore-unsafeinline-expected.txt

CONSOLE MESSAGE: line 17: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
ALERT: PASS (1/2)
ALERT: PASS (2/2)
This tests that a valid nonce disables inline JavaScript, even if 'unsafe-inline' is present.

PASS
